+++ tests/kuku/common.cpp
@@ -4,6 +4,7 @@
 #include "kuku/common.h"
 #include "gtest/gtest.h"
 #include <cmath>
+#include <cstdint>
 
 using namespace kuku;
 using namespace std;
@@ -13,64 +14,68 @@ namespace kuku_tests
     TEST(CommonTests, SetItem)
     {
         item_type bl;
+
         set_item(0, 0, bl);
-        ASSERT_EQ(0, bl[0]);
-        ASSERT_EQ(0, bl[1]);
+        ASSERT_EQ(0, get_low_word(bl));
+        ASSERT_EQ(0, get_high_word(bl));
 
         set_item(1, 0, bl);
-        ASSERT_EQ(1, bl[0]);
-        ASSERT_EQ(0, bl[1]);
+        ASSERT_EQ(1, get_low_word(bl));
+        ASSERT_EQ(0, get_high_word(bl));
 
         set_item(0, 1, bl);
-        ASSERT_EQ(0, bl[0]);
-        ASSERT_EQ(1, bl[1]);
+        ASSERT_EQ(0, get_low_word(bl));
+        ASSERT_EQ(1, get_high_word(bl));
 
         set_item(0xF00F, 0xBABA, bl);
-        ASSERT_EQ(0xF00F, bl[0]);
-        ASSERT_EQ(0xBABA, bl[1]);
+        ASSERT_EQ(0xF00F, get_low_word(bl));
+        ASSERT_EQ(0xBABA, get_high_word(bl));
 
         set_item(0xF00FF00FF00FF00F, 0xBABABABABABABABA, bl);
-        ASSERT_EQ(0xF00FF00FF00FF00F, bl[0]);
-        ASSERT_EQ(0xBABABABABABABABA, bl[1]);
+        ASSERT_EQ(0xF00FF00FF00FF00F, get_low_word(bl));
+        ASSERT_EQ(0xBABABABABABABABA, get_high_word(bl));
 
         unsigned char data[bytes_per_item]{ 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07,
                                             0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07 };
         set_item(data, bl);
-        ASSERT_EQ(0x0706050403020100, bl[0]);
-        ASSERT_EQ(0x0706050403020100, bl[1]);
+        ASSERT_EQ(0x0706050403020100, get_low_word(bl));
+        ASSERT_EQ(0x0706050403020100, get_high_word(bl));
     }
 
     TEST(CommonTests, SetZeroItem)
     {
         item_type bl;
+
         set_item(0, 0, bl);
         set_zero_item(bl);
-        ASSERT_EQ(0, bl[0]);
-        ASSERT_EQ(0, bl[1]);
+        ASSERT_EQ(0, get_low_word(bl));
+        ASSERT_EQ(0, get_high_word(bl));
 
         set_item(0xF00FF00FF00FF00F, 0xBABABABABABABABA, bl);
         set_zero_item(bl);
-        ASSERT_EQ(0, bl[0]);
-        ASSERT_EQ(0, bl[1]);
+        ASSERT_EQ(0, get_low_word(bl));
+        ASSERT_EQ(0, get_high_word(bl));
     }
 
     TEST(CommonTests, SetAllOnesItem)
     {
         item_type bl;
+
         set_zero_item(bl);
         set_all_ones_item(bl);
-        ASSERT_EQ(0xFFFFFFFFFFFFFFFFULL, bl[0]);
-        ASSERT_EQ(0xFFFFFFFFFFFFFFFFULL, bl[1]);
+        ASSERT_EQ(0xFFFFFFFFFFFFFFFFULL, get_low_word(bl));
+        ASSERT_EQ(0xFFFFFFFFFFFFFFFFULL, get_high_word(bl));
 
         set_item(0xF00FF00FF00FF00F, 0xBABABABABABABABA, bl);
         set_all_ones_item(bl);
-        ASSERT_EQ(0xFFFFFFFFFFFFFFFFULL, bl[0]);
-        ASSERT_EQ(0xFFFFFFFFFFFFFFFFULL, bl[1]);
+        ASSERT_EQ(0xFFFFFFFFFFFFFFFFULL, get_low_word(bl));
+        ASSERT_EQ(0xFFFFFFFFFFFFFFFFULL, get_high_word(bl));
     }
 
     TEST(CommonTests, IsZeroItem)
     {
         item_type bl;
+
         set_item(0, 0, bl);
         ASSERT_TRUE(is_zero_item(bl));
 
@@ -87,6 +92,7 @@ namespace kuku_tests
     TEST(CommonTests, IsAllOnesItem)
     {
         item_type bl;
+
         set_all_ones_item(bl);
         ASSERT_TRUE(is_all_ones_item(bl));
 
@@ -103,6 +109,7 @@ namespace kuku_tests
     TEST(CommonTests, SetRandomItem)
     {
         item_type bl;
+
         set_random_item(bl);
         ASSERT_FALSE(is_zero_item(bl));
         item_type bl2 = bl;
@@ -122,24 +129,25 @@ namespace kuku_tests
 
     TEST(CommonTests, IncrementItem)
     {
-        item_type it = make_item(0, 0);
-        increment_item(it);
-        ASSERT_EQ(1, it[0]);
-        ASSERT_EQ(0, it[1]);
+        item_type bl = make_item(0, 0);
 
-        it = make_item(0xF00F, 0xBAAB);
-        increment_item(it);
-        ASSERT_EQ(0xF010, it[0]);
-        ASSERT_EQ(0xBAAB, it[1]);
+        increment_item(bl);
+        ASSERT_EQ(1, get_low_word(bl));
+        ASSERT_EQ(0, get_high_word(bl));
 
-        it = make_item(0xFFFFFFFFFFFFFFFF, 0xFFFFFFFFFFFFFFFE);
-        increment_item(it);
-        ASSERT_EQ(0x0, it[0]);
-        ASSERT_EQ(0xFFFFFFFFFFFFFFFF, it[1]);
+        bl = make_item(0xF00F, 0xBAAB);
+        increment_item(bl);
+        ASSERT_EQ(0xF010, get_low_word(bl));
+        ASSERT_EQ(0xBAAB, get_high_word(bl));
 
-        it = make_item(0xFFFFFFFFFFFFFFFF, 0xFFFFFFFFFFFFFFFF);
-        increment_item(it);
-        ASSERT_EQ(0x0, it[0]);
-        ASSERT_EQ(0x0, it[1]);
+        bl = make_item(0xFFFFFFFFFFFFFFFF, 0xFFFFFFFFFFFFFFFE);
+        increment_item(bl);
+        ASSERT_EQ(0x0, get_low_word(bl));
+        ASSERT_EQ(0xFFFFFFFFFFFFFFFF, get_high_word(bl));
+
+        bl = make_item(0xFFFFFFFFFFFFFFFF, 0xFFFFFFFFFFFFFFFF);
+        increment_item(bl);
+        ASSERT_EQ(0x0, get_low_word(bl));
+        ASSERT_EQ(0x0, get_high_word(bl));
     }
 } // namespace kuku_tests